requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
antlr4-python3-runtime==4.12.0
APScheduler==3.11.0
attrs==24.2.0
babel==2.16.0
bitarray==2.9.3
bitstring==4.2.3
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpickle==3.1.0
frozenlist==1.5.0
holidays==0.62
idna==3.10
kazoo==2.10.0
magic-filter==1.0.12
multidict==6.1.0
overrides==7.7.0
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
pytils==0.4.1
pytz==2024.2
pyutils==0.0.14
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
yarl==1.18.3
python-dotenv~=1.0.1
redis-om~=0.3.3